refactor advisory policy evaluation routes
