+++ src/app/routers/advisory_policy_evaluation_common.py
@@ -1,0 +1,6 @@
+from fastapi import Path
+
+POLICY_EVALUATION_PATH = Path(
+    ...,
+    description="Policy evaluation identifier owned by lotus-advise.",
+)
